+++ CHANGELOG.rst
@@ -6,6 +6,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+[v0.1.2] - 2023-05-15
+---------------------
+
+Added
+~~~~~
+
+* test cases for `get_import_path_for_xml_mapper` function
+
+Fixes
+~~~~~
+
+* fixes wrong version missmatching in `get_import_path_for_xml_mapper` function
+
 [v0.1.1] - 2023-05-15
 ---------------------